pyproject.toml
[project]
name = "devex-worker"
version = "0.1.0"
description = "CLI and execution engine for DevEx workloads"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
